Refactor Attachment into a separate module
